Extract `BranchHeader` from inside `BranchCard`
